fix: load model in FP16 to save ~2 GB VRAM on T4
AutoModelForCausalLM was loaded in float32 by default, consuming
twice as much GPU memory as needed before any quantization started.
Switching to float16 frees ~2 GB of VRAM, which is critical on a
14.5 GB T4 for the 345M GPT-2 medium and 1.24B LLaMA experiments.

+++ experiments/run_experiment.py
@@ -62,7 +62,7 @@ def load_model_and_tokenizer(model_name: str, device: str = "cuda"):
 
     model = AutoModelForCausalLM.from_pretrained(
         model_name,
-        torch_dtype=torch.float32,   # FP32 for precise quantization; baselines cast down as needed
+        torch_dtype=torch.float16,   # FP16 loading saves ~2 GB VRAM on T4; kernels upcast internally
         device_map="auto" if device == "cuda" else None,
         trust_remote_code=True,
     )
